# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPay API",
        "description": "Documentation de l'API OMPay",
        "termsOfService": "https://ompay.com/terms",
        "contact": {
            "email": "contact@ompay.com"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://ompay.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API OMPay Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/v1/comptes": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau client et son compte",
                "operationId": "cbcca7c7331be9c2aa63764eef3dee50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_naissance",
                                    "telephone",
                                    "cni",
                                    "email",
                                    "sexe"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "cni": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "sexe": {
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre compte a été créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Données invalides"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors de la création du compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/comptes/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Consulter le solde d'un compte spécifique",
                "description": "Permet à un utilisateur authentifié de consulter le solde d'un compte spécifique en fournissant son ID",
                "operationId": "f5997e071d30f2fba6973b03ec2130f7",
                "parameters": [
                    {
                        "name": "compte_id",
                        "in": "query",
                        "description": "ID du compte dont on veut consulter le solde",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde du compte récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "C20231112345"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "date_consultation": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "L'ID du compte est requis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé à ce compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/comptes/details": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher les détails d'un compte",
                "operationId": "fc743c3235b4d6ad10ea459afa282a41",
                "parameters": [
                    {
                        "name": "numero_compte",
                        "in": "query",
                        "description": "Numéro de compte",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "compte": {
                                                    "$ref": "#/components/schemas/Compte"
                                                },
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Transaction"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Numéro de compte manquant ou invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le token JWT: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}